refactor: 컴포넌트 이름 변경 - input창: Textarea - 댓글: CommentInput - 답글: ReplyInput (#33)
